change on prob
proba par default + taille par default dans promap

+++ proMap.py
@@ -31,7 +31,7 @@ def reset():
 
 def buttonPlusSimu():
     if not textBoxProba.get() == "":
-        vg.setProba(int(textBoxProba.get()))
+        vg.setProba(float(textBoxProba.get()))
     lancer_chrono()
     sim_auto()
